+++ RELEASES.md
@@ -3,7 +3,8 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 `2.0.9-SNAPSHOT`
 * service-template: returning hash code and location of submitted record
 * service-template: verifying hash from the command line
-
+* prov-dot: allows for dot:style to be specified
+* 
 `2.0.8`
 * module-tools: prov-template-archetype and prov-toolbox-plugin
 * prov-template-compiler: record index with principal